requirements.txt
openai 
fastapi 
uvicorn 
python-dateutil
python-dotenv